requirements.txt
Flask==3.0.0
gevent==23.9.1
keras==2.13.1
numpy==1.24.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
Werkzeug==3.0.0
